moved to to new words generator
